# Switch some low-level logging in hashing related code to 'if' guards