Didn't expect much gain as the arguments involved don't require complex evaluation. However the CONIKS hasher turned out faster by ~10% using this.

Results were repeatable across multiple runs.

Before:

BenchmarkHStar2Root-12    	     200	  47001010 ns/op
BenchmarkHStar2Root-12    	     200	  46416230 ns/op
BenchmarkHStar2Root-12    	     200	  47124380 ns/op

After:

BenchmarkHStar2Root-12    	     200	  41636550 ns/op
BenchmarkHStar2Root-12    	     200	  41859620 ns/op
BenchmarkHStar2Root-12    	     200	  42533060 ns/op

### Checklist

<!---
Go over all the following points, and put an `x` in all the boxes that apply.
Feel free to not tick any boxes that don't apply to this PR (e.g. refactoring may not need a CHANGELOG update).
If you're unsure about any of these, don't hesitate to ask. We're here to help!
-->

- [ ] I have updated the [CHANGELOG](CHANGELOG.md).
  - Adjust the draft version number according to [semantic versioning](https://semver.org/) rules.
- [ ] I have updated [documentation](docs/) accordingly.